+++ Backend/src/Models/GoalsModel.php
@@ -8,7 +8,7 @@ use DateTime;
 
 class GoalsModel
 {
-    public function __construct(private int $id, private string $name, private string $description, private string $status, private string $type, private int $progress,  private DateTime $createdAt, private DateTime $deadline) {}
+    public function __construct(private int $id, private string $name, private string $description, private string $status, private string $type, private float $progress,  private DateTime $createdAt, private DateTime $deadline) {}
     public function getId()
     {
         return $this->id;
